requirements.txt
torchvision
pytorch
matplotlib
scipy
numpy
opencv-python
tqdm
yacs
h5py
scipy
Pillow